test(ssh): replace Tailscale harness with lightweight none-auth tunnel test (#4358)
The #4358 regression is purely client-side: in agent mode with an empty
ssh-agent, the tunnel must still offer the SSH `none` auth method. Any server
that accepts `none` proves this, so the headscale/Tailscale harness is overkill.

Removes the dedicated test:tailscale suite (jest config, scripts, CI job,
headscale compose/policy, testcontainers harness) and adds a normal integration
test, tests/integration/lib/db/clients/ssh-none.spec.ts, modeled on
ssh-agent.spec.js (#4193). It stands up an in-process ssh2 server that accepts
only `none` and forwards its direct-tcpip channel to a postgres container, then
connects in agent mode with a key-less ssh-agent.

Verified RED on the pre-fix ['agent']-only handler ("All configured
authentication methods failed") and GREEN with the 'none' fallback already on
master. Runs in seconds via yarn test:integration; no extra images or CI config.
