+++ apps/studio/package.json
@@ -27,7 +27,6 @@
     "internal:integration": "cross-env TEST_MODE=1 ELECTRON_RUN_AS_NODE=1 yarn electron ../../node_modules/jest/bin/jest.js --config ./jest.integration.config.js",
     "test:integration": "../../bin/integration-tests.sh",
     "test:integration:macos": "cross-env TEST_MODE=1 ELECTRON_RUN_AS_NODE=1 yarn electron ../../node_modules/jest/bin/jest.js --config ./jest.integration.macos.config.js",
-    "test:tailscale": "cross-env TEST_MODE=1 ELECTRON_RUN_AS_NODE=1 yarn electron ../../node_modules/jest/bin/jest.js --config ./jest.tailscale.config.js",
     "dev:esbuild": "node ./esbuild.mjs watch",
     "dev:vite": "vite dev"
   },
