--- apps/studio/tests/docker/headscale/config.yaml
@@ -1,51 +1,0 @@
-# Minimal headscale control-plane config for the tailscale SSH integration test.
-# Bind-mounted into the headscale container at /etc/headscale/config.yaml.
-#
-# Everything stays self-contained: sqlite state, an embedded DERP server so no
-# external relays are needed, and a file-based ACL policy (policy.hujson) that
-# grants Tailscale SSH access. The nodes live on the same docker network and
-# connect directly, so the embedded DERP only exists to satisfy headscale's
-# requirement for at least one DERP region.
-server_url: http://headscale:8080
-listen_addr: 0.0.0.0:8080
-metrics_listen_addr: 0.0.0.0:9090
-
-noise:
-  private_key_path: /var/lib/headscale/noise_private.key
-
-prefixes:
-  v4: 100.64.0.0/10
-  v6: fd7a:115c:a1e0::/48
-  allocation: sequential
-
-derp:
-  server:
-    enabled: true
-    region_id: 999
-    region_code: headscale
-    region_name: Headscale Embedded DERP
-    stun_listen_addr: 0.0.0.0:3478
-    private_key_path: /var/lib/headscale/derp_server_private.key
-  urls: []
-  paths: []
-  auto_update_enabled: false
-  update_frequency: 24h
-
-database:
-  type: sqlite
-  sqlite:
-    path: /var/lib/headscale/db.sqlite
-
-dns:
-  magic_dns: true
-  base_domain: tailnet.test
-  nameservers:
-    global:
-      - 1.1.1.1
-
-policy:
-  mode: file
-  path: /etc/headscale/policy.hujson
-
-log:
-  level: info
