+++ package.json
@@ -23,7 +23,6 @@
     "test:e2e": "yarn workspace beekeeper-studio test:e2e",
     "test:e2e:ci": "yarn workspace beekeeper-studio test:e2e:ci",
     "test:integration": "yarn workspace beekeeper-studio test:integration",
-    "test:tailscale": "yarn workspace beekeeper-studio test:tailscale",
     "test:unit": "yarn workspace beekeeper-studio test:unit && yarn workspace @beekeeperstudio/ui-kit test:unit",
     "test:ci": "yarn workspace beekeeper-studio test:ci",
     "test:codemirror": "yarn workspace beekeeper-studio test:codemirror",
